style(display_controller): widen loop var to match upper bound
clang-tidy flagged bugprone-too-small-loop-variable on
  for (uint8_t i = 0; i < CYCLE_LEN; i++)
where CYCLE_LEN is sizeof()/sizeof() → size_t. CYCLE_LEN is 4 today
so it never overflows uint8_t, but the comparison-with-wider-type
pattern is exactly the kind of thing that bites later if the array
grows. Widen the loop variable to size_t and cast on return (the
returned index fits an array of <256 entries, which the cycle
order is bounded by anyway).

+++ Software/STMCubeV2/BatteryMonitorV2/Core/Src/display_controller.c
@@ -81,9 +81,9 @@ static const osThreadAttr_t s_task_attr = {
 /* Find the index of `view` in kCycleOrder, or 0 if not found. */
 static uint8_t cycle_index_of(display_view_t view)
 {
-    for (uint8_t i = 0; i < CYCLE_LEN; i++) {
+    for (size_t i = 0; i < CYCLE_LEN; i++) {
         if (kCycleOrder[i] == view) {
-            return i;
+            return (uint8_t)i;
         }
     }
     return 0;
